# contracts/openapi.json
{
  "openapi": "3.0.4",
  "info": {
    "title": "Condux Control Plane API",
    "version": "v1"
  },
  "servers": [
    {
      "url": "http://localhost:8099/"
    }
  ],
  "paths": {
...
    "/api/admin/orgs/{orgId}/billing": {
      "get": {
        "tags": [
          "Admin"
        ],
        "operationId": "adminBillingStatus",
        "parameters": [
          {
            "name": "orgId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int64"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AdminBillingResponse"
                }
              }
            }
          },
          "404": {
            "description": "Not Found"
          }
        }
      }
    },
    "/api/admin/orgs/{orgId}/billing/cancel": {
      "post": {
        "tags": [
          "Admin"
        ],
        "operationId": "adminBillingCancel",
        "parameters": [
          {
            "name": "orgId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int64"
            }
          }
        ],
        "responses": {
          "202": {
            "description": "Accepted"
          },
          "404": {
            "description": "Not Found"
          },
          "400": {
            "description": "Bad Request",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      }
    },
    "/api/admin/orgs/{orgId}/billing/plan": {
      "post": {
        "tags": [
          "Admin"
        ],
        "operationId": "adminBillingChangePlan",
        "parameters": [
          {
            "name": "orgId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int64"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/AdminChangePlanRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "202": {
            "description": "Accepted"
          },
          "404": {
            "description": "Not Found"
          },
          "400": {
            "description": "Bad Request",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      }
    },
    "/api/admin/orgs/{orgId}/billing/portal": {
      "post": {
        "tags": [
          "Admin"
        ],
        "operationId": "adminBillingPortal",
        "parameters": [
          {
            "name": "orgId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int64"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AdminPortalResponse"
                }
              }
            }
          },
          "404": {
            "description": "Not Found"
          },
          "400": {
            "description": "Bad Request",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      }
    },
    "/api/admin/spend": {
      "get": {
        "tags": [
          "Admin"
        ],
        "operationId": "adminSpendRollup",
        "parameters": [
          {
            "name": "days",
            "in": "query",
            "schema": {
              "type": "integer",
              "format": "int32",
              "default": 30
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AdminSpendRollupResponse"
                }
              }
            }
          }
        }
      }
    },
    "/api/admin/orgs/{orgId}/spend": {
      "get": {
        "tags": [
          "Admin"
        ],
        "operationId": "adminOrgSpend",
        "parameters": [
          {
            "name": "orgId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int64"
            }
          },
          {
            "name": "days",
            "in": "query",
            "schema": {
              "type": "integer",
              "format": "int32",
              "default": 30
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AdminSpendRollupResponse"
                }
              }
            }
          }
        }
      }
    },
    "/api/admin/spend/runs": {
      "get": {
        "tags": [
          "Admin"
        ],
        "operationId": "adminSpendRuns",
        "parameters": [
          {
            "name": "orgId",
            "in": "query",
            "schema": {
              "type": "integer",
              "format": "int64"
            }
          },
          {
            "name": "kind",
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "days",
            "in": "query",
            "schema": {
              "type": "integer",
              "format": "int32",
              "default": 30
            }
          },
          {
            "name": "limit",
            "in": "query",
            "schema": {
              "type": "integer",
              "format": "int32",
              "default": 200
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/AdminSpendRunResponse"
                  }
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "AdminBillingResponse": {
        "required": [
          "tier",
          "billingEnabled",
          "hasSubscription",
          "stripeCustomerId",
          "stripeSubscriptionId",
          "purchasableTiers"
        ],
        "type": "object",
        "properties": {
          "tier": {
            "type": "string"
          },
          "billingEnabled": {
            "type": "boolean"
          },
          "hasSubscription": {
            "type": "boolean"
          },
          "stripeCustomerId": {
            "type": "string",
            "nullable": true
          },
          "stripeSubscriptionId": {
            "type": "string",
            "nullable": true
          },
          "purchasableTiers": {
            "type": "array",
            "items": {
              "type": "string"
            }
          }
        }
      },
      "AdminChangePlanRequest": {
        "required": [
          "tier"
        ],
        "type": "object",
        "properties": {
          "tier": {
            "type": "string"
          }
        }
      },
...
      "AdminPortalResponse": {
        "required": [
          "url"
        ],
        "type": "object",
        "properties": {
          "url": {
            "type": "string"
          }
        }
      },
      "AdminSpendModelResponse": {
        "required": [
          "model",
          "runCount",
          "inputTokens",
          "outputTokens",
          "costUsd"
        ],
        "type": "object",
        "properties": {
          "model": {
            "type": "string"
          },
          "runCount": {
            "type": "integer",
            "format": "int32"
          },
          "inputTokens": {
            "type": "integer",
            "format": "int64"
          },
          "outputTokens": {
            "type": "integer",
            "format": "int64"
          },
          "costUsd": {
            "type": "number",
            "format": "double",
            "nullable": true
          }
        }
      },
      "AdminSpendOrgResponse": {
        "required": [
          "orgId",
          "orgName",
          "runCount",
          "inputTokens",
          "outputTokens",
          "costUsd"
        ],
        "type": "object",
        "properties": {
          "orgId": {
            "type": "integer",
            "format": "int64"
          },
          "orgName": {
            "type": "string"
          },
          "runCount": {
            "type": "integer",
            "format": "int32"
          },
          "inputTokens": {
            "type": "integer",
            "format": "int64"
          },
          "outputTokens": {
            "type": "integer",
            "format": "int64"
          },
          "costUsd": {
            "type": "number",
            "format": "double",
            "nullable": true
          }
        }
      },
      "AdminSpendRollupResponse": {
        "required": [
          "totalUsd",
          "byModel",
          "byOrg"
        ],
        "type": "object",
        "properties": {
          "totalUsd": {
            "type": "number",
            "format": "double"
          },
          "byModel": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/AdminSpendModelResponse"
            }
          },
          "byOrg": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/AdminSpendOrgResponse"
            }
          }
        }
      },
      "AdminSpendRunResponse": {
        "required": [
          "id",
          "orgId",
          "orgName",
          "model",
          "kind",
          "status",
          "inputTokens",
          "outputTokens",
          "costUsd",
          "createdAt"
        ],
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid"
          },
          "orgId": {
            "type": "integer",
            "format": "int64"
          },
          "orgName": {
            "type": "string"
          },
          "model": {
            "type": "string"
          },
          "kind": {
            "type": "string"
          },
          "status": {
            "type": "integer",
            "format": "int32"
          },
          "inputTokens": {
            "type": "integer",
            "format": "int64"
          },
          "outputTokens": {
            "type": "integer",
            "format": "int64"
          },
          "costUsd": {
            "type": "number",
            "format": "double",
            "nullable": true
          },
          "createdAt": {
            "type": "string",
            "format": "date-time"
          }
        }
      },
...
      "ErrorResponse": {
        "required": [
          "error"
        ],
        "type": "object",
        "properties": {
          "error": {
            "type": "string"
          }
        }
      },
...
    }
  },
...
}